+++ src/include/memory.h
@@ -43,7 +43,7 @@ extern "C" {
 struct stream_buffer
 {
    char* buffer;  /**< allocated buffer holding streaming data */
-   int size;      /**< allocated buffer size */
+   size_t size;      /**< allocated buffer size */
    int start;     /**< offset to the first unconsumed data in buffer */
    int end;       /**< offset to the first position after available data */
    int cursor;    /**< next byte to consume */
